Added a new check to a command that checks if the user is human or a whitelisted bot,
